+++ fonte/ieadam/src/main/java/br/com/ieadam/repositorio/RotinaRepositorioSqlImpl.java
@@ -17,7 +17,7 @@ public class RotinaRepositorioSqlImpl implements RotinaRepositorioSql {
 
 	@SuppressWarnings("unchecked")
 	@Override
-	public List<Rotina> listaRotinasPorPerfil(Long id) {
+	public List<Rotina> listaRotinasPorPerfil(int id) {
 		return entityManager
 				.createNativeQuery(
 						"SELECT i.* FROM saa_rotina i, saa_perfil_rotina b where i.id_rotina = b.id_rotina and id_perfil = " + id,
